# Add faster identifier parsing routines