--- a/2023/09/03/benchmarks/benchmark.cpp
+++ b/2023/09/03/benchmarks/benchmark.cpp
@@ -56,6 +56,24 @@ int main(int argc, char **argv) {
                        exit(1);
                    }
                }));
+  pretty_print(input_data.size(), identifiers, "count_identifiers_neon_strager",
+               bench([&input_data, &sum, &identifiers]() {
+                   size_t r = count_identifiers_neon_strager(input_data.data(), input_data.size());
+                   sum += r;
+                   if(r != identifiers) {
+                       printf("FAIL %zu %zu\n", r, identifiers);
+                       exit(1);
+                   }
+               }));
+  pretty_print(input_data.size(), identifiers, "count_identifiers_neon_strager_ranges",
+               bench([&input_data, &sum, &identifiers]() {
+                   size_t r = count_identifiers_neon_strager_ranges(input_data.data(), input_data.size());
+                   sum += r;
+                   if(r != identifiers) {
+                       printf("FAIL %zu %zu\n", r, identifiers);
+                       exit(1);
+                   }
+               }));
   pretty_print(input_data.size(), identifiers, "count_identifiers",
                bench([&input_data, &sum, &identifiers]() {
                    size_t r = count_identifiers(input_data.data(), input_data.size());
@@ -65,4 +83,4 @@ int main(int argc, char **argv) {
                        exit(1);
                    }
                }));
-  }
+}