--- a/2023/09/03/include/identifiers.h
+++ b/2023/09/03/include/identifiers.h
@@ -7,5 +7,7 @@
 /* Identifiers can start with an ascii letter or _, and they can contain letters, _ or digits. */
 size_t count_identifiers(const char *source, size_t size);
 size_t count_identifiers_neon(const char *source, size_t size);
+size_t count_identifiers_neon_strager(const char *source, size_t size);
+size_t count_identifiers_neon_strager_ranges(const char *source, size_t size);
 
 #endif // NAME_TO_DNSWIRE_H