Implement two improvements to count_identifiers_neon:

* count_identifiers_neon_strager: 5-6% speedup over count_identifiers_neon. Use a faster method of generating the bitmasks. This bitmask generated is 32-bit instead of 16-bit, consuming two bit per uint8x16_t element instead of one bit per, necessitating some changes to the main loop.

* count_identifiers_neon_strager_ranges: 7-8% speedup over count_identifiers_neon. Optimized bitmask as above, and also use ranges ('a' <= c <= 'z', for example) instead of tables for classification.

Performance results on my Apple M1 with Apple clang version 14.0.0 (clang-1400.0.29.202) (fastest of 3):

count_identifiers_neon                   :   0.47 GB/s  8984.6 Ma/s   0.11 ns/d
count_identifiers_neon_strager           :   0.50 GB/s  9497.3 Ma/s   0.11 ns/d
count_identifiers_neon_strager_ranges    :   0.51 GB/s  9662.1 Ma/s   0.10 ns/d
count_identifiers                        :   0.04 GB/s   824.1 Ma/s   1.21 ns/d